requirements.txt
openmetadata-ingestion==1.5.5
requests
python-dotenv